TechFarming/requirements.txt
adafruit-circuitpython-dht
adafruit-blinka
pyserial
RPi.GPIO
requests
